requirements.txt
streamlit==1.39.0
pandas==2.2.3
python-dateutil==2.9.0.post0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.27
langgraph==0.2.3
groq==0.6.0
python-dotenv==1.0.1
tqdm==4.66.5
